Move startup code
