+++ internal/util/startup/startup.go
@@ -12,11 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+// Package startup provides initialization code shared by main and ferretdb packages.
+package startup
 
-import "fmt"
+import (
+	"fmt"
+	"path/filepath"
 
-// stateFileProblem returns the state file access error.
-func stateFileProblem(_ string, err error) string {
-	return fmt.Sprintf("Failed to create state provider: %s.", err)
+	"github.com/FerretDB/FerretDB/v2/internal/util/state"
+)
+
+// State setups state provider for the given directory.
+func State(dir string) (*state.Provider, error) {
+	if dir == "" {
+		return nil, fmt.Errorf("state directory is not set")
+	}
+
+	f, err := filepath.Abs(filepath.Join(dir, "state.json"))
+	if err != nil {
+		return nil, err
+	}
+
+	sp, err := state.NewProvider(f)
+	if err != nil {
+		return nil, stateFileErr(f, err)
+	}
+
+	return sp, nil
 }
